+++ samples/snippets/csharp/concepts/linq/LinqSamples/csv/names.csv
