+++ samples/snippets/csharp/concepts/linq/LinqSamples/csv/scores.csv
